requirements.txt
python-telegram-bot
httpx
openpyxl
Pillow
